converter_agent/requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0
azure-cosmos>=4.2.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-docx>=0.8.11
pdfminer.six>=20221105
